pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "leaf-eater"
authors = [{ name = "David Ward" }, { name = "Matiiss" }]
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["numpy>=2.2.3", "pygame-ce>=2.5.3"]

[project.optional-dependencies]
dev = ["ruff"]

[tool.pylint.main]
disable = ["all"]
enable = ["attribute-defined-outside-init"]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "I", "C901"]

[tool.ruff.lint.isort]
combine-as-imports = true
